Fix 3C test that fails only on Windows.
The test 3C/multiple_tu.c fails only on Windows. The problem is the test script. The
script generates a command file in JSON format by piping a small generated Python
program into the Pyython interpreter.  The program contains strings for file names.
On Windows, the directory separate is a backslash character (`\`). The backslash is interpreted as a string escape character. This results in JSON with incorrect paths.
Use raw Python strings for the file names so that Python leaves them alone.