Fix compiler assert involving a bound-safe interface with a typedef'ed type.
The conversion of a pointer-typed variable with a bounds-safe interface to
an unchecked pointer type was failing when the bound-safe interface type used
a typedef'ed type. There was an assertion that a type had not been
not been canonicalized. The fix is to use the canonicalized version of the
bounds-safe interface type in the conversion code.